files/requirements.txt
torch
transformers
Pillow
timm
flask